Update to M size hmubar

+++ Data.py
@@ -5,6 +5,7 @@ import subprocess
 import sys
 import os
 
+reset = 0
 Folder_malbec = 'Run_New_4/'
 Folder_laptop = 'Run_New_4/'
 Workspace_laptop = 'Workspace_New_4/'
@@ -13,13 +14,14 @@ Root_laptop = '/Users/thorey/Documents/These/Projet/Refroidissement/Skin_Model/S
 Workspace_laptop = '/Users/thorey/Documents/These/Projet/Refroidissement/Skin_Model/SCAPAD/ELAS/'+Workspace_laptop
 Routine_python = '/Users/thorey/Documents/These/Projet/Refroidissement/Skin_Model/Routine/ELAS/E_Load.py'
 
-try :
-    subprocess.call('rm -R '+Root_laptop +'*',shell = True)
-    subprocess.call('rm '+Workspace_laptop+'*',shell =True)
-except:
-    print Root_laptop
-    print 'error'
-    sys.exit()
+if reset == 1:
+    try :
+        subprocess.call('rm -R '+Root_laptop +'*',shell = True)
+        subprocess.call('rm '+Workspace_laptop+'*',shell =True)
+    except:
+        print Root_laptop
+        print 'error'
+        sys.exit()
 try:
     subprocess.call('rsync -zav -e"ssh -p11270" thorey@localhost:' +Root_malbec+ ' ' + Root_laptop, shell = True)
 except:
